requirements.txt
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
tqdm==4.66.4
pillow==10.3.0
Flask==3.0.3
gunicorn==21.2.0
